Fixed array_slice and array_splice to use 64 bit integers for length

+++ src/runtime/base/array/array_util.cpp
@@ -96,7 +96,7 @@ Variant ArrayUtil::Chunk(CArrRef input, int size,
   return ret;
 }
 
-Variant ArrayUtil::Slice(CArrRef input, int offset, int length,
+Variant ArrayUtil::Slice(CArrRef input, int offset, int64 length,
                          bool preserve_keys) {
   int num_in = input.size();
   if (offset > num_in) {
@@ -128,7 +128,7 @@ Variant ArrayUtil::Slice(CArrRef input, int offset, int length,
   return out_hash;
 }
 
-Variant ArrayUtil::Splice(CArrRef input, int offset, int length /* = 0 */,
+Variant ArrayUtil::Splice(CArrRef input, int offset, int64 length /* = 0 */,
                           CVarRef replacement /* = null_variant */,
                           Array *removed /* = NULL */) {
   int num_in = input.size();
